+++ e2e-tests/tests/angular-19.test.ts
@@ -231,6 +231,7 @@ function checkAngularProject(
       'replaysSessionSampleRate: 0.1',
       'replaysOnErrorSampleRate: 1',
       'enableLogs: true',
+      'sendDefaultPii: true',
     ]);
   });